feat: setup esm with TS
