--- foo/bar.ts
@@ -1,1 +1,0 @@
-export function baz(){ console.log("Hello From Baz. ")};
