+++ package.json
@@ -7,27 +7,22 @@
   "license": "MIT",
   "private": false,
   "type": "module",
-  "_moduleAliases": {
-    "@config": "dist/config",
-    "@lib": "dist/lib",
-    "@models": "dist/models",
-    "@plugins": "dist/plugins",
-    "@routes": "dist/routes",
-    "@services": "dist/services",
-    "@api": "dist/api/index"
-  },
   "scripts": {
-    "tsc": "tsc",
-    "start": "tsc && node --experimental-modules ./dist/src/index.js"
-  },
-  "devDependencies": {
-    "dotenv": "^8.2.0",
-    "tslint": "^6.1.3"
+    "watch": "tsc --watch",
+    "dev": "nodemon ./dist/index.js",
+    "serve": "node ./dist/index.js",
+    "start": "concurrently \"yarn:watch\" \"yarn:dev\""
   },
   "dependencies": {
+    "dotenv": "^8.2.0",
     "firebase": "^7.21.0",
-    "module-alias": "^2.2.2",
     "nodemon": "^2.0.4",
     "typescript": "^4.0.3"
+  },
+  "devDependencies": {
+    "@types/node": "^14.11.2",
+    "concurrently": "^5.3.0",
+    "prettier": "^2.1.2",
+    "swc": "^1.0.11"
   }
 }
