+++ src/config/config.ts
@@ -1,0 +1,13 @@
+import * as dotenv from "dotenv"
+
+dotenv.config()
+
+export const FirebaseConfig = {
+  apiKey: "your-api-key",
+  authDomain: "your-auth-domain",
+  databaseURL: "your-db-url",
+  projectId: "your-project-id",
+  storageBucket: "your-storage-bucket",
+  messagingSenderId: "your-sender-id",
+  appId: "your-api-key",
+}
