+++ src/foo/bar.ts
@@ -1,0 +1,3 @@
+export function baz() {
+  console.log("Hello From Baz.")
+}
