+++ src/index.ts
@@ -1,5 +1,3 @@
-// import { FirebaseConfig } from '../config/config';
-import {baz} from '../foo/bar'
- baz();
-import * as firebase from 'firebase/app';
-// const initialize = firebase.initializeApp(FirebaseConfig);
+import { baz } from "./foo/bar.js"
+
+baz()
